requirements.txt
# Core ML and Data Processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
xgboost>=1.6.0
lightgbm>=3.3.0
optuna>=3.0.0

# Deep Learning and NLP
torch>=1.12.0
transformers>=4.20.0
sentence-transformers>=2.2.0
openai>=0.27.0

# API and Web Framework
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
httpx>=0.24.0

# Database and Caching
psycopg2-binary>=2.9.0
redis>=4.0.0
sqlalchemy>=1.4.0

# Monitoring and Logging
prometheus-client>=0.15.0
structlog>=22.0.0
sentry-sdk>=1.20.0

# Data Pipeline
kafka-python>=2.0.0
celery>=5.2.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0.0
click>=8.0.0
rich>=12.0.0
schedule>=1.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
